requirements.txt
Flask==3.0.3
opencv-python-headless==4.10.0.84
numpy==1.26.4
torch
torchvision
Pillow==10.4.0